feat(desktop): add settings button to title bar and remove redundant footer settings
refactor title bar to include a settings button for easier access to settings view. remove footer buttons in home view as settings handling is consolidated in the title bar. enhance maintainability by centralizing settings logic in `TitleBarController`.
