+++ src/main/java/br/com/gzlabs/gzassist/desktop/HomeController.java
@@ -1,20 +1,10 @@
 package br.com.gzlabs.gzassist.desktop;
 
 import br.com.gzlabs.gzassist.core.Mode;
-import br.com.gzlabs.gzassist.util.ThemeManager;
-import javafx.application.Platform;
 import javafx.fxml.FXML;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
 import javafx.scene.control.Button;
-import javafx.stage.Modality;
-import javafx.stage.Stage;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import java.io.IOException;
-import java.util.Objects;
 
 public class HomeController {
 
@@ -34,6 +24,7 @@ public class HomeController {
     private Button autoDetectBtn;
 
     private Button selectedBtn;
+    
 
     @FXML
     protected void onExamQuestionClick() {
@@ -68,33 +59,6 @@ public class HomeController {
         highlightSelected(autoDetectBtn);
         selectedMode = Mode.AUTO_DETECT;
         LOG.info("Mode: AutoDetect (AI decides) selected");
-    }
-
-    @FXML
-    protected void onSettingsClick() {
-        try {
-            FXMLLoader loader = new FXMLLoader(getClass().getResource("/desktop/settings/settings-view.fxml"));
-            Parent root = loader.load();
-
-            Scene scene = new Scene(root, 650, 450);
-            scene.getStylesheets().add(Objects.requireNonNull(getClass().getResource("/desktop/settings/settings.css")).toExternalForm());
-            ThemeManager.applyTheme(scene, root);
-
-            Stage stage = new Stage();
-            stage.setTitle("Settings");
-            stage.setScene(scene);
-            stage.initModality(Modality.APPLICATION_MODAL);
-            stage.initOwner(examQuestionBtn.getScene().getWindow());
-            stage.showAndWait();
-        } catch (IOException e) {
-            LOG.error("Failed to open the settings view.", e);
-        }
-    }
-
-    @FXML
-    protected void onExitClick() {
-        LOG.info("Exit button clicked. Closing application.");
-        Platform.exit();
     }
 
     private void highlightSelected(Button btn) {
