+++ src/main/resources/desktop/common/main-style.css
@@ -75,26 +75,3 @@
     -fx-font-smoothing-type: lcd;
     -fx-padding: 6 0 0 0;
 }
-
-.footer-hbox {
-    -fx-padding: 22 0 60 0;
-}
-
-.footer-btn {
-    -fx-font-size: 14px;
-    -fx-background-radius: 10;
-    -fx-border-radius: 10;
-    -fx-border-width: 1;
-    -fx-padding: 7 22 7 22;
-    -fx-cursor: hand;
-    -fx-transition: background-color 0.18s, -fx-text-fill 0.15s;
-}
-
-.footer-btn:hover {
-    -fx-opacity: 0.9;
-}
-
-.footer-btn:pressed {
-    -fx-scale-x: 0.98;
-    -fx-scale-y: 0.98;
-}
